Struct vfs::file_set::FileSetConfig

pub struct FileSetConfig { /* private fields */ }

This contains path prefixes to partition a Vfs into FileSets.

Example

let mut builder = FileSetConfigBuilder::default();
builder.add_file_set(vec![VfsPath::new_virtual_path("/src".to_string())]);
let config = builder.build();
let mut file_system = Vfs::default();
file_system.set_file_contents(VfsPath::new_virtual_path("/src/main.rs".to_string()), Some(vec![]));
file_system.set_file_contents(VfsPath::new_virtual_path("/src/lib.rs".to_string()), Some(vec![]));
file_system.set_file_contents(VfsPath::new_virtual_path("/build.rs".to_string()), Some(vec![]));
// contains the sets :
// { "/src/main.rs", "/src/lib.rs" }
// { "build.rs" }
let sets = config.partition(&file_system);

Implementations

impl FileSetConfig

pub fn builder() -> FileSetConfigBuilder

Returns a builder for FileSetConfig.

pub fn partition(&self, vfs: &Vfs) -> Vec<FileSet>

Partition vfs into FileSets.

Creates a new FileSet for every set of prefixes in self.

pub fn roots(&self) -> Vec<(Vec<u8>, u64)>

Get the lexicographically ordered vector of the underlying map.