Enum vfs::loader::Message

pub enum Message {
    Progress {
        n_total: usize,
        n_done: Option<usize>,
        dir: Option<AbsPathBuf>,
        config_version: u32,
    },
    Loaded {
        files: Vec<(AbsPathBuf, Option<Vec<u8>>)>,
    },
    Changed {
        files: Vec<(AbsPathBuf, Option<Vec<u8>>)>,
    },
}

Message about an action taken by a Handle.

Variants

Progress

Fields

n_total: usize

The total files to be loaded.

n_done: Option<usize>

The files that have been loaded successfully.

dir: Option<AbsPathBuf>

The dir being loaded, None if its for a file.

config_version: u32

The Config version.

Indicate a gradual progress.

This is supposed to be the number of loaded files.

Loaded

Fields

files: Vec<(AbsPathBuf, Option<Vec<u8>>)>

The handle loaded the following files’ content for the first time.

Changed

Fields

files: Vec<(AbsPathBuf, Option<Vec<u8>>)>

The handle loaded the following files’ content.