vfs
Struct VfsPath

pub struct VfsPath(/* private fields */);

Path in Vfs.

Long-term, we want to support files which do not reside in the file-system, so we treat VfsPaths as opaque identifiers.

Implementations

impl VfsPath

pub fn new_virtual_path(path: String) -> VfsPath

Creates an โin-memoryโ path from /-separated string.

This is most useful for testing, to avoid windows/linux differences

Panics

Panics if path does not start with '/'.

pub fn new_real_path(path: String) -> VfsPath

Create a path from string. Input should be a string representation of an absolute path inside filesystem

pub fn as_path(&self) -> Option<&AbsPath>

Returns the AbsPath representation of self if self is on the file system.

pub fn join(&self, path: &str) -> Option<VfsPath>

Creates a new VfsPath with path adjoined to self.

pub fn pop(&mut self) -> bool

Remove the last component of self if there is one.

If self has no component, returns false; else returns true.

Example

let mut path = VfsPath::from(AbsPathBuf::assert("/foo/bar".into()));
assert!(path.pop());
assert_eq!(path, VfsPath::from(AbsPathBuf::assert("/foo".into())));
assert!(path.pop());
assert_eq!(path, VfsPath::from(AbsPathBuf::assert("/".into())));
assert!(!path.pop());

pub fn starts_with(&self, other: &VfsPath) -> bool

Returns true if other is a prefix of self.

pub fn strip_prefix(&self, other: &VfsPath) -> Option<&RelPath>

pub fn parent(&self) -> Option<VfsPath>

Returns the VfsPath without its final component, if there is one.

Returns None if the path is a root or prefix.

pub fn name_and_extension(&self) -> Option<(&str, Option<&str>)>

Returns selfโs base name and file extension.