vfs
Module file_set

Partitions a list of files into disjoint subsets.

Files which do not belong to any explicitly configured FileSet belong to the default FileSet.

Structs

FileSet

A set of VfsPaths identified by FileIds.

FileSetConfig

This contains path prefixes to partition a Vfs into FileSets.

FileSetConfigBuilder

Builder for FileSetConfig.